pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "numpy>=2.2.6",
    "onnxruntime>=1.22.0",
    "opencv-python>=4.11.0.86",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.5",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "python-telegram-bot>=22.1",
    "rembg>=2.0.66",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
